+++ source/language/en.js
@@ -202,7 +202,9 @@ module.exports = {
     ascmd: {
       help: 'Enter ashelp to view local commands',
       ashelp: `Usage:
- ascmd file\t\tExecute the command with file, eg: ascmd /bin/bash
+ ascmd [file]\t\tExecute the command with file, eg: ascmd /bin/bash
+ aslistcmd\t\tList available command interpreters
+ aspowershell [on|off]\t\tEnable/Disable PowerShell mode, eg: aspowershell on
  quit\t\tClose terminal
  exit\t\tClose terminal
 
@@ -210,9 +212,23 @@ Hot Keys:
  Ctrl =\t\tIncrease font
  Ctrl -\t\tDecrease font
  Ctrl L\t\tClean screen
- Ctrl U\t\tClear the current row
+ Ctrl U\t\tClear the current line
+ Ctrl A\t\tMove cursor to the beginning of the line
+ Ctrl E\t\tMove cursor to the end of the line
+ Ctrl F/B\t\tForward and backward (equivalent to the left and right direction keys)
+ Ctrl P\t\tPrevious command
+ Ctrl R\t\tSearch command history
+ Ctrl D\t\tDelete the character of the current cursor
+ Ctrl H\t\tDeletes the character before the cursor
+ Ctrl W\t\tDelete the word before the cursor
+ Ctrl K\t\tDelete to the end of the text
+ Ctrl T\t\tExchange text at the cursor
  `,
       ascmd: (cmd) => antSword.noxss(`Will execute the command with ${cmd}.`),
+      aspowershell: {
+        on: "Powershell mode enabled",
+        off: "Powershell mode disabled",
+      },
     },
   },
   filemanager: {
@@ -555,7 +571,8 @@ Hot Keys:
       homepage: 'Home',
       document: 'Document',
       qqgroup: 'QQ Group',
-      discord: 'Discord'
+      discord: 'Discord',
+      wechat: 'Fllow us on WeChat'
     },
     language: {
       title: 'Language setting',