+++ source/modules/terminal/index.js
@@ -40,7 +40,9 @@ class Terminal {
     this.term = null;
     this.cell = cell;
     this.isWin = true;
+    this.isPowershell = false;
     this.sessbin = null;
+    this.sess_powershell = null;
     this.core = new antSword['core'][opts['type']](opts);
     this.cache = new antSword['CacheManager'](this.opts['_id']);
 
@@ -150,6 +152,54 @@ class Terminal {
         term.echo(LANG['ascmd']['ashelp']);
         return;
       }
+      if (cmd === 'aslistcmd'){
+        var binarr = "";
+        if (this.isWin) {
+          binarr = [
+            "C:/Windows/System32/cmd.exe",
+            "C:/Windows/SysWOW64/cmd.exe",
+            "C:/Windows/System32/WindowsPowerShell/v1.0/powershell.exe",
+            "C:/Windows/SysWOW64/WindowsPowerShell/v1.0/powershell.exe",
+            "C:/Windows/System32/WindowsPowerShell/v2.0/powershell.exe",
+            "C:/Windows/SysWOW64/WindowsPowerShell/v2.0/powershell.exe",
+            "C:/Windows/System32/WindowsPowerShell/v3.0/powershell.exe",
+            "C:/Windows/SysWOW64/WindowsPowerShell/v3.0/powershell.exe",
+            "C:/Windows/System32/WindowsPowerShell/v4.0/powershell.exe",
+            "C:/Windows/SysWOW64/WindowsPowerShell/v4.0/powershell.exe",
+          ].join(',');
+        }else{
+          binarr = [
+            "/bin/sh",
+            "/bin/ash",
+            "/bin/bash",
+            "/bin/zsh",
+            "/bin/busybox",
+          ].join(',');
+        }
+        this.core.request(
+          this.core.command.listcmd({
+            binarr: binarr,
+          })
+        ).then((ret) => {
+          let res = ret['text'];
+          if(res.indexOf("ERROR://") > -1){
+            throw res;
+          }
+          let result = "";
+          res.split('\n').map((v) => {
+            var line = v.split('\t');
+            if(line.length == 2){
+              var r = parseInt(line[1]) === 1 ? '[[b;#15af63;]OK]' : '[[b;#E80000;]FAIL]';
+              result += `${line[0]}\t\t\t${r}\n`;
+            }
+          });
+          term.echo(result);
+          term.resume();
+        }).catch((err) => {
+          term.resume();
+        });
+        return;
+      }
       if ( cmd.substr(0,5) === 'ascmd') {
         var sessbin = cmd.substr(5).trim();
         if(sessbin.length>0){
@@ -157,6 +207,17 @@ class Terminal {
           term.echo(LANG['ascmd']['ascmd'](self.sessbin));
         }else{
           term.echo(LANG['ascmd']['ashelp']);
+        }
+        return;
+      }
+      if ( cmd.substr(0,12) === 'aspowershell') {
+        var _switch = cmd.substr(12).trim().toLowerCase();
+        if(_switch === "on") {
+          self.sess_powershell = true;
+          term.echo(LANG['ascmd']['aspowershell']["on"]);
+        }else {
+          self.sess_powershell = false;
+          term.echo(LANG['ascmd']['aspowershell']["off"]);
         }
         return;
       }
@@ -178,6 +239,14 @@ class Terminal {
       _bin = _confBin || _bin;
       if(self.sessbin !== null) {
         _bin = self.sessbin;
+      }
+      if(self.isWin && _bin.indexOf("powershell") > -1) {
+        self.isPowershell = true
+      }else{
+        self.isPowershell = false
+      }
+      if(self.sess_powershell !== null) {
+        self.isPowershell = self.sess_powershell;
       }
       // 开始执行命令
       this.core.request(
@@ -230,7 +299,7 @@ class Terminal {
         // < 1.0.0 时使用3个参数 completion: (term, value, callback) => {}
         completion: (value, callback) => {
           callback([
-            'ashelp', 'ascmd', 'quit', 'exit'
+            'ashelp', 'ascmd', 'aslistcmd', 'aspowershell', 'quit', 'exit'
           ].concat(
             this.isWin ? [
               'dir', 'whoami', 'net', 'ipconfig', 'netstat', 'cls',
@@ -286,7 +355,7 @@ class Terminal {
   parseCmd(cmd, path) {
     path = path.replace(/\\\\/g, '\\').replace(/"/g, '\\"').replace(/\\/g, '\\\\');
     return (this.isWin
-      ? `cd /d "${path}"&${cmd}&echo [S]&cd&echo [E]`
+      ? this.isPowershell? `cd "${path}";${cmd};echo [S];(pwd).path;echo [E]`:`cd /d "${path}"&${cmd}&echo [S]&cd&echo [E]`
       : `cd "${path}";${cmd};echo [S];pwd;echo [E]`
     );
   }